requirements/common.txt
django==2.1.2
psycopg2==2.5.4
whitenoise==2.0.6
django-environ==0.4.0
djangorestframework==3.9.0
requests
Pillow